config/settings.py
from pathlib import Path
import os
import importlib.util

from dotenv import load_dotenv


BASE_DIR = Path(__file__).resolve().parent.parent

load_dotenv(BASE_DIR / ".env")


def _env_bool(name: str, default: bool = False) -> bool:
    value = os.getenv(name)
    if value is None:
        return default
    normalized = value.strip().lower()
    if normalized in {"1", "true", "yes", "on"}:
        return True
    if normalized in {"0", "false", "no", "off"}:
        return False
    return default


def _env_int(name: str, default: int) -> int:
    value = os.getenv(name)
    if value is None:
        return default
    try:
        return int(value)
    except (TypeError, ValueError):
        return default


SECRET_KEY = os.getenv("SECRET_KEY", "django-insecure-dev-secret-key")
DEBUG = _env_bool("DEBUG", default=True)
ALLOWED_HOSTS = [host.strip() for host in os.getenv("ALLOWED_HOSTS", "127.0.0.1,localhost").split(",") if host.strip()]

INSTALLED_APPS = [
    "django.contrib.admin",
    "django.contrib.auth",
    "django.contrib.contenttypes",
    "django.contrib.sessions",
    "django.contrib.messages",
    "django.contrib.staticfiles",
    "apps.tracks",
    "apps.users",
    "apps.interactions",
]

if importlib.util.find_spec("jazzmin") is not None:
    INSTALLED_APPS.insert(0, "jazzmin")

MIDDLEWARE = [
    "django.middleware.security.SecurityMiddleware",
    "django.contrib.sessions.middleware.SessionMiddleware",
    "django.middleware.locale.LocaleMiddleware",
    "django.middleware.common.CommonMiddleware",
    "django.middleware.csrf.CsrfViewMiddleware",
    "django.contrib.auth.middleware.AuthenticationMiddleware",
    "django.contrib.messages.middleware.MessageMiddleware",
    "django.middleware.clickjacking.XFrameOptionsMiddleware",
]

ROOT_URLCONF = "config.urls"

TEMPLATES = [
    {
        "BACKEND": "django.template.backends.django.DjangoTemplates",
        "DIRS": [BASE_DIR / "templates"],
        "APP_DIRS": True,
        "OPTIONS": {
            "context_processors": [
                "django.template.context_processors.debug",
                "django.template.context_processors.request",
                "django.template.context_processors.static",
                "django.template.context_processors.media",
                "django.template.context_processors.tz",
                "django.contrib.auth.context_processors.auth",
                "django.contrib.messages.context_processors.messages",
            ],
        },
    },
]

WSGI_APPLICATION = "config.wsgi.application"
ASGI_APPLICATION = "config.asgi.application"

DATABASES = {
    "default": {
        "ENGINE": "django.db.backends.sqlite3",
        "NAME": BASE_DIR / "db.sqlite3",
    }
}

AUTH_PASSWORD_VALIDATORS = [
    {"NAME": "django.contrib.auth.password_validation.UserAttributeSimilarityValidator"},
    {"NAME": "django.contrib.auth.password_validation.MinimumLengthValidator"},
    {"NAME": "django.contrib.auth.password_validation.CommonPasswordValidator"},
    {"NAME": "django.contrib.auth.password_validation.NumericPasswordValidator"},
]

LANGUAGE_CODE = "kk"
LANGUAGES = [
    ("kk", "Қазақша"),
    ("ru", "Русский"),
]
LOCALE_PATHS = [BASE_DIR / "locale"]
TIME_ZONE = "UTC"
USE_I18N = True
USE_TZ = True

STATIC_URL = "/static/"
STATICFILES_DIRS = [BASE_DIR / "static"]

MEDIA_URL = "/media/"
MEDIA_ROOT = BASE_DIR / "media"
ENABLE_YTDLP_YOUTUBE_IMPORT = DEBUG and _env_bool("ENABLE_YTDLP_YOUTUBE_IMPORT", default=True)
YTDLP_IMPORT_AUDIO_DIR = BASE_DIR / os.getenv("YTDLP_IMPORT_AUDIO_DIR", "media/audio")
YTDLP_MAX_FILENAME_LENGTH = _env_int("YTDLP_MAX_FILENAME_LENGTH", default=120)

LOGIN_URL = "users:login"
LOGIN_REDIRECT_URL = "tracks:home"
LOGOUT_REDIRECT_URL = "tracks:home"

JAZZMIN_SETTINGS = {
    "site_title": "QazSound Control Panel",
    "site_header": "QazSound Admin",
    "site_brand": "QazSound",
    "site_logo": "img/placeholders/avatar.png",
    "site_icon": "img/placeholders/avatar.png",
    "login_logo": "img/placeholders/avatar.png",
    "welcome_sign": "Welcome to QazSound Control Panel",
    "copyright": "QazSound",
    "search_model": ["tracks.Track", "tracks.Artist", "auth.User", "interactions.Like"],
    "topmenu_links": [
        {"name": "Dashboard", "url": "admin:index", "permissions": ["auth.view_user"]},
        {"model": "tracks.Track"},
        {"model": "tracks.Artist"},
        {"model": "interactions.Like"},
    ],
    "show_sidebar": True,
    "navigation_expanded": False,
    "hide_apps": ["contenttypes", "sessions"],
    "icons": {
        "auth": "fas fa-users-cog",
        "auth.user": "fas fa-user",
        "tracks": "fas fa-music",
        "tracks.track": "fas fa-wave-square",
        "tracks.artist": "fas fa-microphone",
        "tracks.genre": "fas fa-tags",
        "interactions": "fas fa-heart",
        "interactions.like": "fas fa-heart",
        "interactions.playlist": "fas fa-list-music",
        "interactions.playlistitem": "fas fa-stream",
    },
    "default_icon_parents": "fas fa-folder",
    "default_icon_children": "fas fa-circle",
    "custom_css": "admin/custom.css",
    "show_ui_builder": False,
    "related_modal_active": True,
}

JAZZMIN_UI_TWEAKS = {
    "navbar_small_text": False,
    "footer_small_text": False,
    "body_small_text": False,
    "brand_small_text": False,
    "brand_colour": "navbar-dark",
    "accent": "accent-info",
    "navbar": "navbar-dark",
    "no_navbar_border": False,
    "navbar_fixed": True,
    "layout_boxed": False,
    "footer_fixed": False,
    "sidebar_fixed": True,
    "sidebar": "sidebar-dark-primary",
    "sidebar_nav_small_text": False,
    "sidebar_disable_expand": False,
    "sidebar_nav_child_indent": False,
    "sidebar_nav_compact_style": True,
    "sidebar_nav_legacy_style": False,
    "sidebar_nav_flat_style": False,
    "theme": "darkly",
    "dark_mode_theme": "darkly",
    "button_classes": {
        "primary": "btn-info",
        "secondary": "btn-outline-secondary",
        "info": "btn-info",
        "warning": "btn-warning",
        "danger": "btn-danger",
        "success": "btn-success",
    },
}

DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
