requirements.txt
django
pillow
python-dotenv
django-jazzmin
yt-dlp
